restructure Firefox code so that it doesn't live in a separate dir
Jetpack allows us to use any directory in our add-on in place of lib/main.js, so I've moved all of that code into background/firefox-main.js.

However, Jetpack still requires content scripts/assets to be in /data. So I scrapped the original /Firefox directory, and moved the original /Firefox/data dir to a top level /data directory, and updated the symlinks accordingly.